Tuple.java update
Passes all the test except modifyRecordID, which the lab syllabus says
we don't need to yet.

I had to make TDItem class in TupleDesc to also implement Field, which
required me to add new methods within it - they're left blank though.
This is because the TDItems needed to be cast as Field - check the
constructor for the tuple